Moved files into more logical directory structure.
