+++ Python/simpledominion/game/card/GameCard.py
@@ -1,8 +1,8 @@
 from typing import Type
-from simpledominion.GameCardType import GameCardType
-from simpledominion.CardInterface import CardInterface
-from simpledominion.GameCardType import GameCardType
-from simpledominion.TurnStatus import TurnStatus
+from simpledominion.game.card.GameCardType import GameCardType
+from simpledominion.game.card.CardInterface import CardInterface
+from simpledominion.game.card.GameCardType import GameCardType
+from simpledominion.game.TurnStatus import TurnStatus
 
 class GameCard(CardInterface):
