+++ Python/simpledominion/game/card/GameCardType.py
