+++ Python/test/test_deck.py
@@ -1,8 +1,8 @@
 from typing import List
 from unittest import TestCase
-from simpledominion.Deck import DeckFactory, DeckInterface
-from simpledominion.GameCardType import GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER, GAME_CARD_TYPE_FESTIVAL
-from simpledominion.CardInterface import CardInterface
+from simpledominion.game.piles.Deck import DeckFactory, DeckInterface
+from simpledominion.game.card.GameCardType import GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER, GAME_CARD_TYPE_FESTIVAL
+from simpledominion.game.card.CardInterface import CardInterface
 from test.fake_card import FakeCard
 
 class MockDiscardPile:
