+++ Python/test/test_discardpile.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
-from simpledominion.DiscardPile import DiscardPileFactory
-from simpledominion.GameCardType import GameCardType, GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER
-from simpledominion.CardInterface import CardInterface
+from simpledominion.game.piles.DiscardPile import DiscardPileFactory
+from simpledominion.game.card.GameCardType import GameCardType, GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER
+from simpledominion.game.card.CardInterface import CardInterface
 
 class FakeCard(CardInterface):
     def __init__(self, cardType: GameCardType):
