+++ Python/test/test_hand.py
@@ -1,7 +1,7 @@
 from typing import List
 from unittest import TestCase
-from simpledominion.Hand import HandFactory, HandInterface
-from simpledominion.GameCardType import GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER, GAME_CARD_TYPE_FESTIVAL
+from simpledominion.game.player.Hand import HandFactory, HandInterface
+from simpledominion.game.card.GameCardType import GAME_CARD_TYPE_ESTATE, GAME_CARD_TYPE_COPPER, GAME_CARD_TYPE_FESTIVAL
 from test.fake_card import FakeCard
 
 class FakeDeck:
